+++ src/Bridge/bridgeSensorsMessages.ts
@@ -1,0 +1,5 @@
+export const ProximitySensorChange = 'Sensors.Proximity.Change';
+export interface ProximitySensorChange {
+	type: typeof ProximitySensorChange;
+	detected: boolean;
+}