+++ src/Driver/Sensors/NotImplementedSensors.ts
@@ -1,0 +1,16 @@
+import ISensors, { SensorCapability } from './ISensors';
+import IProximitySensor from '@signageos/front-display/es6/NativeDevice/Sensors/IProximitySensor';
+import NotImplementedProximitySensor from '@signageos/front-display/es6/Sensors/NotImplementedProximitySensor';
+
+export default class NotImplementedSensors implements ISensors {
+
+	public readonly proximity: IProximitySensor;
+
+	constructor() {
+		this.proximity = new NotImplementedProximitySensor();
+	}
+
+	public supports(_capability: SensorCapability): boolean {
+		return false;
+	}
+}